Moving some c# code around
