+++ src/api/PatientAdd.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.Collections.Generic;
+using Contoso.Healthcare.Api.Models;
 
 namespace Contoso
 {
